Avoid hexadecimal floats
The cs2cs utility fails to interpret some hexadecimal float values (but works just fine for many others). The idea was to send data to cs2cs without loss of precision, but in practice that precision probably isn't even required most of the time.

On a plate carrée, a decimal number with 11 significant digits provides millimetre precision for the entire world. Even accounting for distortion due to the projection in use, 12 significant digits should really be enough for basically every use case.

Therefore, FORMAT_OUT is well-served with %.12g.

Since the base of an IEEE double has approx. 16 decimal significant digits, a sensible FORMAT_IN would seem to be %.15g. It avoids the hex issues, is easy to parse, and provides sufficient, but not pathologically excessive precision. It does provide a bit of "extra" precision to increase the chance of numerical stability during round-trip (though this depends primarily on PROJ's math).

This change also adds a test for some floats that are known to be problematic when %a is used with cs2cs.