Move package to Geo::LibProj namespace
Using the Geo::Proj5 namespace didn't really make sense because this module is not limited to any particular PROJ version.

In the abstract, it would make sense for this module to live in the Geo::PROJ namespace. However, that namespace collides with the unrelated Geo::Proj module on case-insensitive file systems.

The Geo::Proj4 project plans to use the Geo::Proj5 namespace for a future variant of Geo::Proj4 that works with PROJ 5. The plan seems to be to keep Geo::Proj as "the abstract interface" (see Geo::Proj4 GH issue 1).

It's important to understand, however, that Geo::Proj isn't an interface to the PROJ library specifically. Instead, it's mainly an abstraction of spatial reference systems of Geo::Point instances that (currently) uses PROJ to do all the math. Therefore, any future module that does in fact represent an abstract interface to the PROJ library, free of historic Geo::Point and Geo::Proj baggage, should probably use another namespace.

I expect Geo::LibProj to be a suitable name for such a generic module. This change keeps that name available for future use.